Add warning that user should check currency symbols mean the same on both exchanges
After getting a message on telegram:
Heads up everyone! I'm new to the autocoin-trader.com site, so forgive me if this is public knowledge already:
The arbitrage-monitor page reports a price discrepancy for NFT/USDT between Kucoin and Bibox.
I just bought some NFT on Kucoin and tried sending it to Bibox and it vanished!
Bibox support pointed out to me that the NFT I sent (on TRC20) is different than the NFT on their exchange (ERC20).
I don't know enough about this stuff to understand how it's possible for a transaction can be completed despite all that... but there it is...
Thankfully I didn't use a lot of money... and hopefully no one else is as stupid as I am...